v0.1.11
- Fixed #23
- Split the item and block enum for clarity
- Added the ForceGenerateChunk function, which syncronously generates chunks on whichever thread called it. This ensures that the chunk definitely exists after it's been called.
- Removed the requirement for a valid spawn point to have Bedrock under it
- Removed inventory defines from player class, moved into inventory header
- Moved labels into their own file, separate from items
- Added WindowItems to the Respond/Responses class, potentially allowing for full inventory updates (needed when opening e.g. chests, furnaces), although this is currently unused and untested.
