+++ src/enums/inventory.h
@@ -1,0 +1,30 @@
+#define INVENTORY_CRAFTING_RESULT 0
+#define INVENTORY_CRAFTING 1
+
+#define INVENTORY_HOTBAR 36
+#define INVENTORY_HOTBAR_LAST 44
+
+#define INVENTORY_HELMET 5
+#define INVENTORY_CHESTPLATE 6
+#define INVENTORY_LEGGINGS 7
+#define INVENTORY_BOOTS 8
+
+#define INVENTORY_ROW_1 9
+#define INVENTORY_ROW_2 18
+#define INVENTORY_ROW_3 27
+#define INVENTORY_ROW_LAST 35
+
+#define INVENTORY_MAX_SLOTS 45
+
+#define INVENTORY_CHEST 0
+#define INVENTORY_WORKBENCH 1
+#define INVENTORY_FURNACE 2
+#define INVENTORY_DISPENSER 3
+
+#define EQUIPMENT_SLOT_HELD 0
+#define EQUIPMENT_SLOT_HELMET 1
+#define EQUIPMENT_SLOT_CHESTPLATE 2
+#define EQUIPMENT_SLOT_LEGGINGS 3
+#define EQUIPMENT_SLOT_BOOTS 4
+
+#define CLICK_OUTSIDE -999
