+++ src/misc/version.h
@@ -2,9 +2,9 @@
 #pragma once
 #define PROJECT_VERSION_MAJOR 0
 #define PROJECT_VERSION_MINOR 1
-#define PROJECT_VERSION_PATCH 10
-#define PROJECT_VERSION_STRING "0.1.10"
+#define PROJECT_VERSION_PATCH 11
+#define PROJECT_VERSION_STRING "0.1.11"
 
 #define PROJECT_NAME "BetrockServer"
 
-#define PROJECT_NAME_VERSION "BetrockServer 0.1.10"
+#define PROJECT_NAME_VERSION "BetrockServer 0.1.11"
