+++ src/objects/world.h
@@ -7,7 +7,7 @@
 #include <filesystem>
 
 #include "helper.h"
-#include "items.h"
+#include "blocks.h"
 #include "generator.h"
 #include "config.h"
